+++ docs/changelog.md
@@ -5,6 +5,8 @@
 * A new config flag `launchAdditionalService` has been added. True by default, it controls the launch of services
 additional service like Forkmon, Prometheus and Grafana, as well as the transaction spammer. Is set to false, only the
 Ethereum network (EL _and_ CL nodes) are launched
+* The testnet-verifier has been removed, and with it its associated config params: `waitForFinalization`, 
+`waitForVerifications` and `verificationsEpochLimit`
 
 ### Changes
 * Updated flags